Enabled serving of sphinx build files in DEBUG mode only
