+++ docs/urls.py
@@ -1,7 +1,8 @@
+from django.conf import settings
 from django.urls import path, re_path
 from django.views.generic import RedirectView
 
-from . import views
+from . import views, views_debug
 
 urlpatterns = [
     path("", views.index, name="homepage"),
@@ -20,29 +21,6 @@ urlpatterns = [
         views.document,
         {"url": ""},
         name="document-index",
-    ),
-    re_path(
-        r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/_objects/$",
-        views.objects_inventory,
-        name="objects-inv",
-    ),
-    re_path(
-        r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/_images/(?P<path>.*)$",
-        views.sphinx_static,
-        {"subpath": "_images"},
-        name="sphinx-images",
-    ),
-    re_path(
-        r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/_source/(?P<path>.*)$",
-        views.sphinx_static,
-        {"subpath": "_sources"},
-        name="sphinx-sources",
-    ),
-    re_path(
-        r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/_downloads/(?P<path>.*)$",
-        views.sphinx_static,
-        {"subpath": "_downloads"},
-        name="sphinx-downloads",
     ),
     re_path("^(.*)/index/$", views.redirect_index),
     re_path(
@@ -65,5 +43,20 @@ urlpatterns = [
         views.document,
         name="document-detail",
     ),
-    re_path(r"^pots/(?P<pot_name>\w+\.pot)$", views.pot_file),
 ]
+
+if settings.DEBUG:
+    # Patterns for sphinx (in production they are served by nginx directly)
+    # They need to be inserted at the beginning to take precedence over document-detail
+    urlpatterns = [
+        re_path(
+            r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/_objects/$",
+            views_debug.objects_inventory,
+        ),
+        re_path(
+            r"^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/(?P<subpath>_downloads|_images|_source)/(?P<path>.*)$",  # noqa: E501
+            views_debug.sphinx_static,
+        ),
+        re_path(r"^pots/(?P<pot_name>\w+\.pot)$", views.pot_file),
+        *urlpatterns,
+    ]
