+++ docs/views.py
@@ -10,7 +10,6 @@ from django.http import Http404, JsonResponse
 from django.shortcuts import redirect, render
 from django.template.response import TemplateResponse
 from django.utils.translation import activate, gettext_lazy as _
-from django.views import static
 from django.views.decorators.cache import cache_page
 from django_hosts.resolvers import reverse
 
@@ -109,30 +108,6 @@ if not settings.DEBUG:
     document = cache_page(settings.CACHE_MIDDLEWARE_SECONDS, cache="docs-pages")(
         document
     )
-
-
-def pot_file(request, pot_name):
-    version = DocumentRelease.objects.current().version
-    doc_root = str(get_doc_root_or_404("en", version, subroot="gettext"))
-    return static.serve(request, document_root=doc_root, path=pot_name)
-
-
-def sphinx_static(request, lang, version, path, subpath=None):
-    """
-    Serve Sphinx static assets from a subdir of the build location.
-    """
-    document_root = str(get_doc_root_or_404(lang, version).joinpath(subpath))
-    return static.serve(request, document_root=document_root, path=path)
-
-
-def objects_inventory(request, lang, version):
-    response = static.serve(
-        request,
-        document_root=str(get_doc_root_or_404(lang, version)),
-        path="objects.inv",
-    )
-    response["Content-Type"] = "text/plain"
-    return response
 
 
 def redirect_index(request, *args, **kwargs):
